requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
flake8==7.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
isort==5.13.2
jiter==0.7.1
lxml==5.3.0
mccabe==0.7.0
openai==1.55.1
pycodestyle==2.12.1
pydantic==2.10.1
pydantic_core==2.27.1
pyflakes==3.2.0
python-docx==1.1.2
python-dotenv==1.0.1
python-telegram-bot==21.7
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
